style(tests): 清掉 clippy 的 unnecessary_cast 与 contains 警告
CI 以 -D warnings 跑 clippy --all-targets, 这三处测试代码会让它红。

+++ wind_input/crates/wind-store/tests/counts_match_listings.rs
@@ -84,7 +84,7 @@ fn the_streaming_walk_sees_exactly_what_the_listing_does() {
         .map(|i| WordIo {
             code: format!("a{i:03}"),
             text: format!("词{i:03}"),
-            weight: (i % 7) as i32,
+            weight: i % 7,
             count: 0,
             boundary: None,
         })
